video: limit the resolution for fixed bimap font
As grub uses fixed bitmap font and also its size is a fixed property, it is not
possible to accommodate to all resolutions, therefore we raise some limit to
the preferred resolution as most themes are designed on popular device and the
resolution in its prime, which is supposedly Full HD.

This change also makes grub font readable on hiDPI device without going through
the steps in.

https://wiki.archlinux.org/index.php/HiDPI#GRUB

v2: efi_gop: Avoid high resolution when trying to keep current mode.