efi: Always try to allocate heap size of 1.6GB
By default, GRUB will allocate a quarter of the pages it got available
in the EFI subsystem. On many current systems, this will amount to
roughly 800MB of RAM assuming an address space of 32 bits. This is
plenty for most use cases, but it doesn't suffice when using full disk
encryption with a key derival function based on Argon2.

Besides the usual iteration count known from PBKDF2, Argon2 introduces
two additional parameters "memory" and "parallelism". While the latter
doesn't really matter to us, the memory parameter is quite interesting.
If encrypting a partition with LUKS2 using Argon2 as KDF, then
cryptsetup will default to a memory parameter of 1GB. Meaning we need to
allocate a buffer of 1GB in size in order to be able to derive the key,
which definitely won't squeeze into the limit of 800MB.

To prepare for Argon2, this commit reworks the memory allocation
algorithm for EFI platforms. Instead of trying to allocate a quarter of
memory available, let's instead introduce a constant target amount of
bytes that we try to allocate. The target is set to the previous value
of MAX_HEAP_SIZE, which amounts to 1.6GB and thus sufficiently high to
accomodate for both Argon2 as well as other functionality. The value is
then clamped to at most half of available memory but at least 100MB.

Link: https://marc.info/?l=grub-devel&m=158386672608347&w=3

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Petr Vorel <petr.vorel@gmail.com>