Patch grub2-efi-xen-removable.patch
bsc#1085842

The grub can be installed with removable option to support booting from
removable media with standard UEFI default file path of the form:
 \EFI\BOOT\BOOT{machine type short-name}.EFI

It does not make use of distributor directory, which becomes a problem for UEFI
Xen installation as it requires that directory to be present for storing xen
stuff like chainloaded hypervisor, xen kernel and so on. Moreover it makes bad
assumption that hypervisor will be chainloaded by grub under the same
directory, which is also not always true.

This patch fixes the problem by ensuring the directory available to Xen
installation if any Xen hypervisor found and independent to grub boot path
$cmdpath to work.