autogen.sh: Detect python
It help to avoid error on distros which has only python3 binary:
./autogen.sh: line 20: python: command not found

Use python3 as the default as python2 is EOL since Jan 2020, but check
also python which is on most distros if not all python2 because
code still works on python2.

Although it should not be needed keep the possibility to define PYTHON.

For detection use "command -v" which is POSIX [3] and supported on all
common shells (bash, zsh, dash, busybox sh, mksh) instead requiring
"which" as extra dependency (usable on containers).

Update INSTALL.

Signed-off-by: Petr Vorel <pvorel@suse.cz>